refactor(mac): extract Vibes trigger actions
- move quick-tap timing, pending tap scheduling, and double-tap Vibe cycling out of TranscriptionManager
- add MacVibesTriggerActionController to own Vibe apply/undo dispatch and pill feedback
- keep TranscriptionManager focused on recording state while delegating tap actions to the Vibes lane
- register the new Mac Vibes controller in the project
- update Mac codemap and engineering docs for the new ownership boundary
